# docs/swagger.yaml
basePath: /
definitions:
  handler.ResponseSuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.errorResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
  handler.getAllListsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.HomeFull'
        type: array
    type: object
...
  model.HomeFull:
    properties:
      id:
        type: integer
      post_body:
        type: string
      post_date:
        type: string
      post_img_path:
        type: string
      post_img_url:
        type: string
      post_title:
        type: string
    type: object
...
  model.NewsPost:
    properties:
      post_body:
        type: string
      post_img_url:
        type: string
      post_title:
        type: string
    type: object
...
  model.TablePost:
    properties:
      post_body:
        type: string
      post_img_url:
        type: string
      post_title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Mr Bobur
  description: API Server for Norbekov App
  title: Norbekov API
  version: "1.0"
paths:
...
  /api/table/create:
    post:
      consumes:
      - application/json
      description: create table_post
      operationId: create-table_post
      parameters:
      - description: table info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.TablePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create TablePost
      tags:
      - Table
  /api/table/delete:
    delete:
      consumes:
      - application/json
      description: delete table
      operationId: delete-table
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllListsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Delete table
      tags:
      - Table
  /api/table/get:
    get:
      consumes:
      - application/json
      description: get table post
      operationId: get-table
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllListsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get table
      tags:
      - Table
  /api/table/update/{id}:
    put:
      consumes:
      - application/json
      description: Update Table by id
      operationId: update-table-id
      parameters:
      - description: Param ID
        in: path
        name: id
        required: true
        type: integer
      - description: home info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.NewsPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update  Table By ID
      tags:
      - Table
  /api/table/upload-img/{id}:
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Upload Table Image
      operationId: upload_img_table
      parameters:
      - description: Param ID
        in: path
        name: id
        required: true
        type: integer
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      - application/octet-stream
      - image/png
      - image/jpeg
      - image/jpg
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload Table Image
      tags:
      - Table
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"